Move parse folder function into `scanner` pkg
